pyproject.toml
[project]
name = "acl-automated-workflow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jinja2>=3.1.6",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
]
